Converted RSS routing controller to TypeScript
no ref

- Modernized ghost/core/core/frontend/services/routing/controllers/rss.js to TypeScript (rss.ts).
- Added type definitions for request options, response contexts, and data objects.
- Updated controller exports in index.js while maintaining backward compatibility.
