+++ ghost/core/core/frontend/services/routing/controllers/index.js
@@ -8,7 +8,7 @@ module.exports = {
     },
 
     get rss() {
-        return require('./rss');
+        return require('./rss').rssController || require('./rss');
     },
 
     get previews() {
